0909_인벤토리 구현 중.. ItemManager.cs 생성 (정적 생성자로 호출됨) 이 곳에 모든 아이템들 생성자로 호출되고, Inventory.cs에서 아이템 추가, 제거 로직 변경

+++ Assets/02_Scripts/Common/UI/FOV PixelPerfect.cs
@@ -3,8 +3,8 @@ using UnityEngine.U2D;
 
 public class FOV_PixelPerfect : MonoBehaviour
 {
-    private const int minPPU = 60;
-    private const int maxPPU = 120;
+    private const byte minPPU = 60;
+    private const byte maxPPU = 120;
 
     public void FOV_Up()
     {
